Add a new component for a pop-out 'meatball' menu
Which will allow per item (row) specific actions
such as grouping things like jumping to a specific tab
in the DetailForm, and if the displayed model is Versioned
then showing relevant (contextual) actions for Publishing,
Unpublishing, and Archiving (deleting) an item. This saves
a user from having to navigate through depths for each
record they wish to perform an action on, without crowding
the actions column for a row on the GridField.